requirements.txt
streamlit
matplotlib
seaborn
pandas
plotly
streamlit-folium
geopy~=2.2.0
leafmap~=0.8.2
osmnx~=1.1.2
networkx~=2.7.1
scikit-learn
altair<5.0.0
streamlit-authenticator
streamlit-awesome-table
streamlit-extras